refactor action switch to be more flat
